// packages/browser/src/generated/api.ts
export type BeginUserRegistration200Options = { [key: string]: unknown };

export type BeginUserRegistration200 = {
  options: BeginUserRegistration200Options;
  userId: string;
};

export type BeginUserRegistration400 = {
  error: string;
};

export type BeginUserRegistration401 = {
  error: string;
};

export type BeginUserRegistration500 = {
  error: string;
};

export type CompleteUserRegistrationBodyCredential = { [key: string]: unknown };

export type CompleteUserRegistrationBody = {
  credential: CompleteUserRegistrationBodyCredential;
  userId: string;
};

export type CompleteUserRegistration200AuthenticationToken = {
  expiresAt: number;
  token: string;
};

export type CompleteUserRegistration200Credential = {
  authenticatorType: string | null;
  id: string;
  label: string | null;
  userId: string;
};

export type CompleteUserRegistration200 = {
  authenticationToken: CompleteUserRegistration200AuthenticationToken;
  credential: CompleteUserRegistration200Credential;
  success: boolean;
  userId: string;
};

export type CompleteUserRegistration400 = {
  error: string;
};

export type CompleteUserRegistration401 = {
  error: string;
};

export type CompleteUserRegistration500 = {
  error: string;
};

export type BeginCredentialRegistrationBody = {
  authenticationToken: string;
  userName?: string;
};

export type BeginCredentialRegistration200Options = { [key: string]: unknown };

export type BeginCredentialRegistration200 = {
  options: BeginCredentialRegistration200Options;
};
// ...
export type CompleteCredentialRegistration200AuthenticationToken = {
  expiresAt: number;
  token: string;
};

export type CompleteCredentialRegistration200Credential = {
  authenticatorType: string | null;
  id: string;
  label: string | null;
  userId: string;
};

export type CompleteCredentialRegistration200 = {
  authenticationToken: CompleteCredentialRegistration200AuthenticationToken;
  credential: CompleteCredentialRegistration200Credential;
  success: boolean;
};

export type CompleteCredentialRegistration400 = {
  error: string;
};

export type CompleteCredentialRegistration401 = {
  error: string;
};

export type CompleteCredentialRegistration500 = {
  error: string;
};

export type UpdateCredentialLabelBody = {
  authenticationToken: string;
  label: string;
};

export type UpdateCredentialLabel400 = {
  error: string;
};

export type UpdateCredentialLabel401 = {
  error: string;
};

export type UpdateCredentialLabel500 = {
  error: string;
};

export type BeginAuthenticationBodyUserIdentifier = {
  userId?: string;
  userName?: string;
};

export type BeginAuthenticationBody = {
  userIdentifier?: BeginAuthenticationBodyUserIdentifier;
};

export type BeginAuthentication200LoginSession = {
  expiresAt: string;
  id: string;
};

export type BeginAuthentication200Options = { [key: string]: unknown };

export type BeginAuthentication200 = {
  loginSession: BeginAuthentication200LoginSession;
  options: BeginAuthentication200Options;
  projectId: string;
};

export type BeginAuthentication400 = {
  error: string;
};

export type CompleteAuthenticationBodyAuthenticationResponse = { [key: string]: unknown };

export type CompleteAuthenticationBody = {
  authenticationResponse: CompleteAuthenticationBodyAuthenticationResponse;
  loginSessionId: string;
};

export type CompleteAuthentication200AuthenticationToken = {
  expiresAt: number;
  token: string;
};

export type CompleteAuthentication200 = {
  authenticationToken: CompleteAuthentication200AuthenticationToken;
};